+++ R/degree.collect.R
@@ -1,17 +1,17 @@
-#' Degree Distribution
+#' Degrees of nodes
 #'
-#' @description Obtain the degrees for all nodes.
-#' @param net	The input network.
+#' @description Collect the degrees for all nodes in a network.
+#' @param net	 The input network.
 #' @details Obtain the degrees for all nodes.
 #' @return A vector.
 #' @author Xu Dong, Nazrul Shaikh.
 #' @examples \dontrun{
 #  x <-  net.erdos.renyi.gnp(1000, 0.01)
-#' x.dist.deg <- dist.deg(x)
-#' summary(x.dist.deg)}
+#' x.deg <- degree.collect(x)
+#' summary(x.deg)}
 #' @export
 
-dist.deg <- function(net){
+degree.collect <- function(net){
 
   lengths(net)