+++ R/neighbors.R
@@ -1,19 +1,16 @@
-#' neighbors of an agent in a network
+#' Neighbors of an agent in a network
 #'
-#' @description Present all neighbors of a give agent in a network.
-#' @param net	The input network.
-#' @param node The input node.
+#' @description Presents all neighbors of a given node.
+#' @param net	 The input network.
+#' @param NodeID The ID of the input node.
 #' @details Neighbors of a node are nodes that directly connects to this node.
 #' @return A vector.
-#' @author Xu Dong, Nazrul Shaikh.
+#' @author Xu Dong
 #' @examples \dontrun{
 #' x <-  net.ring.lattice(12,4)
 #' neighbors(x,2)}
 #' @export
 
-neighbors <- function(net,node) {
-
-
-  net[[node]]
-
+neighbors <- function(net,NodeID) {
+  net[[NodeID]]
 }